rag-application/backend/rag-politecnica-function/requirements.txt
azure-functions
langchain
langchain-openai
azure-identity
azure-keyvault-secrets
azure-search-documents